Move schema functions to /common, add constants
